requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
dotenv-python==0.0.1
Flask==1.1.1
Flask-Cors==3.0.8
geographiclib==1.50
geopy==1.21.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
